+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 * Optimized email CSS
+* Booking confirmation email now has a link to access barcode
 
 ### 1.6.1: June 22th, 2021
 * Remove error logs